consumer/app/pyproject.toml
[tool.poetry]
name = "consumer"
version = "0.1.0"
description = ""
authors = ["Sam <xneklyudx@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
aiokafka = "^0.7.2"
tortoise-orm = "^0.19.2"
aerich = "^0.6.3"
betterconf = "^2.6.1"
asyncpg = "^0.26.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
